fix: Move over styled and usetheme to common
