+++ modules/_labs/core/react/stories/stories_theme.tsx
@@ -7,7 +7,8 @@ import withReadme from 'storybook-readme/with-readme';
 import CanvasProvider from '../lib/CanvasProvider';
 import {CanvasTheme, CanvasThemePalette, Themeable, createCanvasTheme} from '../lib/theming';
 import README from '../lib/theming/README.md';
-import {H1, colors, type, spacing, borderRadius, useTheme} from '@workday/canvas-kit-react-core';
+import {H1, colors, type, spacing, borderRadius} from '@workday/canvas-kit-react-core';
+import {useTheme} from '@workday/canvas-kit-react-common';
 
 const Palettes = styled('div')({
   display: 'flex',
