+++ modules/button/react/lib/DeleteButton.tsx
@@ -1,9 +1,8 @@
 import * as React from 'react';
 import {Themeable, CanvasTheme} from '@workday/canvas-kit-labs-react-core';
 import {ButtonColors, ButtonSize} from './types';
-import {useTheme} from '@workday/canvas-kit-react-core';
 import {ButtonContainer, ButtonLabel} from './parts';
-import {GrowthBehavior} from '@workday/canvas-kit-react-common';
+import {GrowthBehavior, useTheme} from '@workday/canvas-kit-react-common';
 
 export interface DeleteButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement>,
