+++ modules/button/react/lib/DropdownButton.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import {Themeable} from '@workday/canvas-kit-labs-react-core';
-import {useTheme} from '@workday/canvas-kit-react-core';
 import {caretDownIcon} from '@workday/canvas-system-icons-web';
-import {GrowthBehavior} from '@workday/canvas-kit-react-common';
+import {GrowthBehavior, useTheme} from '@workday/canvas-kit-react-common';
 import {DropdownButtonVariant, ButtonIconPosition} from './types';
 import {ButtonContainer, ButtonLabel, ButtonLabelIcon} from './parts';
 import {getButtonColors} from './Button';
