+++ modules/button/react/lib/TextButton.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {Themeable, CanvasTheme, type} from '@workday/canvas-kit-labs-react-core';
-import {focusRing} from '@workday/canvas-kit-react-common';
-import {colors, spacing, borderRadius, useTheme} from '@workday/canvas-kit-react-core';
+import {focusRing, useTheme} from '@workday/canvas-kit-react-common';
+import {colors, spacing, borderRadius} from '@workday/canvas-kit-react-core';
 import {CanvasSystemIcon} from '@workday/design-assets-types';
 import {TextButtonVariant, ButtonIconPosition, ButtonColors} from './types';
 import {ButtonContainer, ButtonLabelIcon, ButtonLabel} from './parts';
