+++ modules/button/react/lib/parts/ButtonLabelData.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {styled} from '@workday/canvas-kit-react-core';
+import {styled} from '@workday/canvas-kit-react-common';
 
 export const buttonLabelDataClassName = 'button-label-data';
