+++ modules/color-picker/react/lib/ColorInput.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {Themeable, ContentDirection} from '@workday/canvas-kit-labs-react-core';
-import {expandHex, GrowthBehavior, ErrorType} from '@workday/canvas-kit-react-common';
-import {colors, spacing, type, inputColors, styled} from '@workday/canvas-kit-react-core';
+import {expandHex, GrowthBehavior, ErrorType, styled} from '@workday/canvas-kit-react-common';
+import {colors, spacing, type, inputColors} from '@workday/canvas-kit-react-core';
 import TextInput, {TextInputProps} from '@workday/canvas-kit-react-text-input';
 import {ColorSwatch} from './parts/ColorSwatch';
