+++ modules/common/react/index.ts
@@ -10,3 +10,5 @@ export * from './lib/parts/_brand-assets';
 export * from './lib/types';
 export * from './lib/genericStyles';
 export * from './lib/Popper';
+export {default as useTheme} from './lib/theming/useTheme';
+export {default as styled} from './lib/theming/styled';
