+++ modules/common/react/lib/styles/focusRing.ts
@@ -1,6 +1,7 @@
 import {keyframes, CSSObject} from '@emotion/core';
-import canvas, {useTheme} from '@workday/canvas-kit-react-core';
+import canvas from '@workday/canvas-kit-react-core';
 import {CanvasTheme} from '@workday/canvas-kit-labs-react-core';
+import {useTheme} from '@workday/canvas-kit-react-common';
 import memoize from 'lodash/memoize';
 
 function calculateFocusRing({
