+++ modules/radio/react/lib/RadioGroup.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import {Themeable} from '@workday/canvas-kit-labs-react-core';
 import Radio, {RadioProps} from './Radio';
-import {borderRadius, spacing, styled} from '@workday/canvas-kit-react-core';
-import {ErrorType, GrowthBehavior, getErrorColors} from '@workday/canvas-kit-react-common';
+import {borderRadius, spacing} from '@workday/canvas-kit-react-core';
+import {ErrorType, GrowthBehavior, getErrorColors, styled} from '@workday/canvas-kit-react-common';
 
 export interface RadioGroupProps extends Themeable, GrowthBehavior {
   /**
