+++ modules/segmented-control/react/lib/SegmentedControl.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import {borderRadius, colors, styled} from '@workday/canvas-kit-react-core';
+import {borderRadius, colors} from '@workday/canvas-kit-react-core';
 import {IconButton, IconButtonProps} from '@workday/canvas-kit-react-button';
-import {mouseFocusBehavior} from '@workday/canvas-kit-react-common';
+import {mouseFocusBehavior, styled} from '@workday/canvas-kit-react-common';
 
 export interface SegmentedControlProps {
   /**
