+++ modules/text-input/react/lib/TextInput.tsx
@@ -1,13 +1,7 @@
 import * as React from 'react';
 import {Themeable} from '@workday/canvas-kit-labs-react-core';
-import {GrowthBehavior, ErrorType, errorRing} from '@workday/canvas-kit-react-common';
-import {
-  borderRadius,
-  inputColors,
-  spacingNumbers,
-  type,
-  styled,
-} from '@workday/canvas-kit-react-core';
+import {GrowthBehavior, ErrorType, errorRing, styled} from '@workday/canvas-kit-react-common';
+import {borderRadius, inputColors, spacingNumbers, type} from '@workday/canvas-kit-react-core';
 
 export interface TextInputProps
   extends Themeable,
